Generalise custom kernels tests
