+++ python/dolfinx/fem/petsc.py
@@ -15,12 +15,8 @@ from __future__ import annotations
 
 import contextlib
 import functools
-import os
-import pathlib
 import typing
 
-import petsc4py
-import petsc4py.lib
 from petsc4py import PETSc
 
 import numpy as np
@@ -44,8 +40,7 @@ __all__ = ["create_vector", "create_vector_block", "create_vector_nest",
            "assemble_vector", "assemble_vector_nest", "assemble_vector_block",
            "assemble_matrix", "assemble_matrix_nest", "assemble_matrix_block",
            "apply_lifting", "apply_lifting_nest", "set_bc", "set_bc_nest",
-           "LinearProblem", "NonlinearProblem",
-           "cffi_utils", "numba_utils", "ctypes_utils"]
+           "LinearProblem", "NonlinearProblem"]
 
 
 def _extract_function_spaces(a: typing.List[typing.List[Form]]):
@@ -778,214 +773,3 @@ class NonlinearProblem:
         A.zeroEntries()
         assemble_matrix_mat(A, self._a, self.bcs)
         A.assemble()
-
-
-def get_petsc_lib() -> pathlib.Path:
-    """Determine path and name of the PETSc library.
-
-    Returns:
-        Full path to the PETSc shared library.
-
-    Raises:
-        RuntimeError: If PETSc library cannot be found for if more than
-            one library is found.
-    """
-    petsc_dir = petsc4py.get_config()['PETSC_DIR']
-    petsc_arch = petsc4py.lib.getPathArchPETSc()[1]
-    candidate_paths = [os.path.join(petsc_dir, petsc_arch, "lib", "libpetsc.so"),
-                       os.path.join(petsc_dir, petsc_arch, "lib", "libpetsc.dylib")]
-    exists_paths = []
-    for candidate_path in candidate_paths:
-        if os.path.exists(candidate_path):
-            exists_paths.append(candidate_path)
-
-    if len(exists_paths) == 0:
-        raise RuntimeError("Could not find a PETSc shared library.")
-    elif len(exists_paths) > 1:
-        raise RuntimeError("More than one PETSc shared library found.")
-
-    return pathlib.Path(exists_paths[0])
-
-
-def load_petsc_lib(loader: typing.Callable[[str], typing.Any]) -> typing.Any:
-    """Load PETSc shared library using loader callable, e.g. ctypes.CDLL.
-
-    Args:
-        loader: A callable that accepts a library path and returns a wrapped library.
-
-    Returns:
-        A wrapped library of the type returned by the callable.
-    """
-    lib_path = get_petsc_lib()
-    try:
-        try:
-            petsc_lib = loader(lib_path)
-        except TypeError:
-            petsc_lib = loader(str(lib_path))
-    except OSError as e:
-        print(f"Failed to load shared library found at {lib_path}.")
-        raise e
-
-    return petsc_lib
-
-
-class numba_utils:
-    """Utility attributes for working with Numba and PETSc.
-
-    These attributes are convenience functions for calling PETSc C
-    functions from within Numba functions.
-
-    Note:
-        `Numba <https://numba.pydata.org/>`_ must be available
-        to use these utilities.
-
-    Examples:
-        A typical use of these utility functions is::
-
-            import numpy as np
-            import numpy.typing as npt
-            def set_vals(A: int,
-                         m: int, rows: npt.NDArray[PETSc.IntType],
-                         n: int, cols: npt.NDArray[PETSc.IntType],
-                         data: npt.NDArray[PETSc.ScalarTYpe], mode: int):
-                MatSetValuesLocal(A, m, rows.ctypes, n, cols.ctypes, data.ctypes, mode)
-    """
-    try:
-        import numba as _numba
-        import llvmlite as _llvmlite
-        _llvmlite.binding.load_library_permanently(str(get_petsc_lib()))
-
-        _int = _numba.from_dtype(PETSc.IntType)
-        _scalar = _numba.from_dtype(PETSc.ScalarType)
-        _real = _numba.from_dtype(PETSc.RealType)
-        _int_ptr = _numba.core.types.CPointer(_int)
-        _scalar_ptr = _numba.core.types.CPointer(_scalar)
-        _MatSetValues_sig = _numba.core.typing.signature(_numba.core.types.intc,
-                                                         _numba.core.types.uintp, _int, _int_ptr,
-                                                         _int, _int_ptr, _scalar_ptr, _numba.core.types.intc)
-        MatSetValuesLocal = _numba.core.types.ExternalFunction("MatSetValuesLocal", _MatSetValues_sig)
-        """See PETSc `MatSetValuesLocal
-        <https://petsc.org/release/manualpages/Mat/MatSetValuesLocal>`_
-        documentation."""
-
-        MatSetValuesBlockedLocal = _numba.core.types.ExternalFunction("MatSetValuesBlockedLocal", _MatSetValues_sig)
-        """See PETSc `MatSetValuesBlockedLocal
-        <https://petsc.org/release/manualpages/Mat/MatSetValuesBlockedLocal>`_
-        documentation."""
-    except ImportError:
-        pass
-
-
-class ctypes_utils:
-    """Utility attributes for working with ctypes and PETSc.
-
-    These attributes are convenience functions for calling PETSc C
-    functions, typically from within Numba functions.
-
-    Examples:
-        A typical use of these utility functions is::
-
-            import numpy as np
-            import numpy.typing as npt
-            def set_vals(A: int,
-                         m: int, rows: npt.NDArray[PETSc.IntType],
-                         n: int, cols: npt.NDArray[PETSc.IntType],
-                         data: npt.NDArray[PETSc.ScalarTYpe], mode: int):
-                MatSetValuesLocal(A, m, rows.ctypes, n, cols.ctypes, data.ctypes, mode)
-    """
-    import ctypes as _ctypes
-    _lib_ctypes = load_petsc_lib(_ctypes.cdll.LoadLibrary)
-
-    # Note: ctypes does not have complex types, hence we use void* for
-    # scalar data
-    _int = np.ctypeslib.as_ctypes_type(PETSc.IntType)
-
-    MatSetValuesLocal = _lib_ctypes.MatSetValuesLocal
-    """See PETSc `MatSetValuesLocal
-    <https://petsc.org/release/manualpages/Mat/MatSetValuesLocal>`_
-    documentation."""
-    MatSetValuesLocal.argtypes = [_ctypes.c_void_p, _int, _ctypes.POINTER(_int), _int,
-                                  _ctypes.POINTER(_int), _ctypes.c_void_p, _ctypes.c_int]
-
-    MatSetValuesBlockedLocal = _lib_ctypes.MatSetValuesBlockedLocal
-    """See PETSc `MatSetValuesBlockedLocal
-    <https://petsc.org/release/manualpages/Mat/MatSetValuesBlockedLocal>`_
-    documentation."""
-    MatSetValuesBlockedLocal.argtypes = [_ctypes.c_void_p, _int, _ctypes.POINTER(_int), _int,
-                                         _ctypes.POINTER(_int), _ctypes.c_void_p, _ctypes.c_int]
-
-
-class cffi_utils:
-    """Utility attributes for working with CFFI (ABI mode) and PETSc.
-
-    These attributes are convenience functions for calling PETSc C
-    functions, typically from within Numba functions.
-
-    Note:
-        `CFFI <https://cffi.readthedocs.io/>`_ and  `Numba
-        <https://numba.pydata.org/>`_ must be available to use these
-        utilities.
-
-    Examples:
-        A typical use of these utility functions is::
-
-            import numpy as np
-            import numpy.typing as npt
-            def set_vals(A: int,
-                         m: int, rows: npt.NDArray[PETSc.IntType],
-                         n: int, cols: npt.NDArray[PETSc.IntType],
-                         data: npt.NDArray[PETSc.ScalarTYpe], mode: int):
-                MatSetValuesLocal(A, m, ffi.from_buffer(rows), n, ffi.from_buffer(cols),
-                                ffi.from_buffer(rows(data), mode)
-    """
-    try:
-        import cffi as _cffi
-        import numba as _numba
-        import numba.core.typing.cffi_utils as _cffi_support
-
-        # Register complex types
-        _ffi = _cffi.FFI()
-        _cffi_support.register_type(_ffi.typeof('float _Complex'), _numba.types.complex64)
-        _cffi_support.register_type(_ffi.typeof('double _Complex'), _numba.types.complex128)
-
-        _lib_cffi = load_petsc_lib(_ffi.dlopen)
-
-        def _petsc_c_types() -> str:
-            assert PETSc.IntType == np.int32 or PETSc.IntType == np.int64
-            if PETSc.IntType == np.int32:
-                c_int_t = "int32_t"
-            elif PETSc.IntType == np.int64:
-                c_int_t = "int64_t"
-
-            scalar_t = PETSc.ScalarType
-            assert scalar_t == np.float32 or scalar_t == np.float64 \
-                or scalar_t == np.complex64 or scalar_t == np.complex128
-            if scalar_t == np.float32:
-                c_scalar_t = "float"
-            elif scalar_t == np.float64:
-                c_scalar_t = "double"
-            elif scalar_t == np.complex64:
-                c_scalar_t = "float _Complex"
-            elif scalar_t == np.complex128:
-                c_scalar_t = "double _Complex"
-            return c_int_t, c_scalar_t
-
-        _c_int_t, _c_scalar_t = _petsc_c_types()
-        _ffi.cdef(f"""
-                int MatSetValuesLocal(void* mat, {_c_int_t} nrow, const {_c_int_t}* irow,
-                                    {_c_int_t} ncol, const {_c_int_t}* icol,
-                                    const {_c_scalar_t}* y, int addv);
-                int MatSetValuesBlockedLocal(void* mat, {_c_int_t} nrow, const {_c_int_t}* irow,
-                                    {_c_int_t} ncol, const {_c_int_t}* icol,
-                                    const {_c_scalar_t}* y, int addv);
-                                    """)
-        MatSetValuesLocal = _lib_cffi.MatSetValuesLocal
-        """See PETSc `MatSetValuesLocal
-        <https://petsc.org/release/manualpages/Mat/MatSetValuesLocal>`_
-        documentation."""
-        MatSetValuesBlockedLocal = _lib_cffi.MatSetValuesBlockedLocal
-        """See PETSc `MatSetValuesBlockedLocal
-        <https://petsc.org/release/manualpages/Mat/MatSetValuesBlockedLocal>`_
-        documentation."""
-    except ImportError:
-        pass
